- Renamed LagomConfig to Config - Renamed experiment_config to config - Made the configuration object not mandatory in experiment.lagom()
