+++ maggy/core/executors/base_executor.py
@@ -15,12 +15,12 @@
 #
 
 from typing import Callable, Any
-from maggy.experiment_config import LagomConfig
+from maggy.config import Config
 
 
 def base_executor_fn(
     train_fn: Callable,
-    config: LagomConfig,
+    config: Config,
 ) -> Callable:
     """Wraps the user supplied training function in order to be passed to the Spark Executors.
