+++ maggy/core/experiment_driver/ablation_driver.py
@@ -18,7 +18,7 @@ import json
 from typing import Callable, Union, Optional
 
 from maggy import util
-from maggy.experiment_config import AblationConfig
+from maggy.config import AblationConfig
 from maggy.earlystop import NoStoppingRule
 from maggy.ablation.ablationstudy import AblationStudy
 from maggy.ablation.ablator.loco import LOCO
