+++ maggy/core/experiment_driver/torch_distributed_training_driver.py
@@ -19,7 +19,7 @@ from typing import Callable, Type, Any
 
 from maggy import util
 from maggy.core.environment.singleton import EnvSing
-from maggy.experiment_config import TorchDistributedConfig
+from maggy.config import TorchDistributedConfig
 from maggy.core.rpc import DistributedTrainingServer
 from maggy.core.experiment_driver.spark_driver import Driver
 from maggy.core.executors.torch_dist_executor import torch_dist_executor_fn
