+++ maggy/experiment/experiment.py
@@ -18,20 +18,27 @@ from maggy.experiment import experiment_python
 from maggy.experiment import experiment_pyspark
 from maggy.core import config as maggyconfig
 from typing import Callable
-from maggy.experiment_config import LagomConfig
+from maggy.config import Config
 
 
-def lagom(train_fn: Callable, config: LagomConfig) -> dict:
+def lagom(train_fn: Callable, config: Config = None) -> dict:
 
     """Entry point for Maggy experiment, this function passes the parameters to the lagom function
     depending whether the kernel is pyspark or python.
     **lagom** is a Swedish word meaning "just the right amount".
 
     :param train_fn: User defined experiment containing the model training.
-    :param config: An experiment configuration. For more information, see experiment_config.
+    :param config: An experiment configuration. For more information, see config.
 
     :returns: The experiment results as a dict.
     """
+
+    if config is None:
+        config = Config(
+            name="maggy_experiment",
+            description="experiment without config object",
+            hb_interval=None,
+        )
     if maggyconfig.is_spark_available():
         return experiment_pyspark.lagom(train_fn, config)
     else:
